requirements.txt
openai
azure-identity
python-dotenv
mcp==1.9.2
termcolor==2.4.0